v2.13.4

+++ installer/windows/installer/installer/blocks/files.cpp
@@ -1,10 +1,12 @@
 #include "files.h"
 
+#include <QStringList>
+
 #include <filesystem>
 #include <shlobj_core.h>
 #include <spdlog/spdlog.h>
 
-#include "../installer_base.h"
+#include "installerenums.h"
 #include "../settings.h"
 #include "../../../utils/applicationinfo.h"
 #include "../../../utils/archive.h"
@@ -46,14 +48,14 @@ int Files::executeStep()
         auto result = ::SHCreateDirectoryEx(NULL, installPath_.c_str(), NULL);
         if (result != ERROR_SUCCESS) {
             spdlog::error(L"Failed to create default install directory ({})", result);
-            return -ERROR_OTHER;
+            return -wsl::ERROR_OTHER;
         }
     }
 
     const wstring exePath = Utils::getExePath();
     if (exePath.empty()) {
         spdlog::error(L"Could not get exe path");
-        return -ERROR_OTHER;
+        return -wsl::ERROR_OTHER;
     }
 
     wsl::Archive archive;
@@ -62,12 +64,12 @@ int Files::executeStep()
     });
 
     if (!archive.extract(L"Windscribe", L"windscribe.7z", exePath, installPath_)) {
-        return -ERROR_OTHER;
+        return -wsl::ERROR_OTHER;
     }
 
     if (!copyLibs()) {
         spdlog::error(L"Failed to copy libs");
-        return -ERROR_OTHER;
+        return -wsl::ERROR_OTHER;
     }
 
     return moveFiles();
@@ -75,6 +77,7 @@ int Files::executeStep()
 
 int Files::moveFiles()
 {
+    int result = -wsl::ERROR_MOVE_CUSTOM_DIR;
     const wstring& settingsInstallPath = Settings::instance().getPath();
 
     try {
@@ -88,7 +91,13 @@ int Files::moveFiles()
         // The target folder, if it exists, is expected to be empty (i.e. we do not allow installing the app into a folder already
         // containing other files).  RemoveDirectory will fail if the target folder is not empty.
         if (filesystem::exists(settingsInstallPath)) {
+            if (!filesystem::is_empty(settingsInstallPath)) {
+                logCustomFolderContents(settingsInstallPath);
+                result = -wsl::ERROR_CUSTOM_DIR_NOT_EMPTY;
+                throw system_error(ERROR_DIR_NOT_EMPTY, generic_category(), "target folder is not empty");
+            }
             if (::RemoveDirectory(settingsInstallPath.c_str()) == FALSE) {
+                result = -wsl::ERROR_DELETE_CUSTOM_DIR;
                 throw system_error(::GetLastError(), generic_category(), "could not remove target folder");
             }
         }
@@ -105,10 +114,12 @@ int Files::moveFiles()
 
         // MoveFileEx cannot be performed atomically if the source and target folders are on different volumes.
         // Thus, the installer only permits installation to a folder on the system drive.
-        BOOL result = ::MoveFileEx(installPath_.c_str(), settingsInstallPath.c_str(), MOVEFILE_WRITE_THROUGH);
-        if (result == FALSE) {
+        BOOL filesMoved = ::MoveFileEx(installPath_.c_str(), settingsInstallPath.c_str(), MOVEFILE_WRITE_THROUGH);
+        if (filesMoved == FALSE) {
             throw system_error(::GetLastError(), generic_category(), "could not move the installed files");
         }
+
+        result = 100;
     }
     catch (system_error& ex) {
         spdlog::error("Could not move installed files: {}", ex.what());
@@ -128,12 +139,11 @@ int Files::moveFiles()
         };
         int ret = SHFileOperation(&fileOp);
         if (ret) {
-            spdlog::error(L"Could not delete partial install: {}", ret);
+            spdlog::error(L"Could not delete partial install from default folder: {}", ret);
         }
-        return -ERROR_MOVE_CUSTOM_DIR;
     }
 
-    return 100;
+    return result;
 }
 
 bool Files::copyLibs()
@@ -176,4 +186,21 @@ bool Files::copyLibs()
     }
 
     return true;
+}
+
+void Files::logCustomFolderContents(const wstring &folder)
+{
+    QStringList sl;
+    error_code ec;
+    for (const auto &entry : filesystem::recursive_directory_iterator(folder, filesystem::directory_options::skip_permission_denied, ec)) {
+        sl.append(QString::fromStdWString(std::filesystem::relative(entry.path(), folder)));
+    }
+
+    if (ec) {
+        spdlog::error("Files::logCustomFolderContents: filesystem::recursive_directory_iterator failed ({})", ec.message().c_str());
+        return;
+    }
+
+    QString fileList = (sl.empty() ? "** no files found **" : sl.join(", "));
+    spdlog::error(L"The custom install folder is not empty.  It contains the following files: {}", fileList.toStdWString());
 }
